app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.2",
    "uvicorn>=0.34.2",
]
